+++ app/code/ThirdParty/BlogArticle/Model/Data/Post.php
@@ -1,0 +1,124 @@
+<?php
+namespace ThirdParty\BlogArticle\Model\Data;
+
+use Magento\Framework\DataObject;
+use ThirdParty\BlogArticle\Api\Data\PostInterface;
+
+class Post extends DataObject implements PostInterface
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function getPostId()
+    {
+        return $this->getData(self::POST_ID) !== null
+            ? (int) $this->getData(self::POST_ID)
+            : null;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setPostId($postId)
+    {
+        return $this->setData(self::POST_ID, $postId);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getTitle()
+    {
+        return $this->getData(self::TITLE);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setTitle($title)
+    {
+        return $this->setData(self::TITLE, $title);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getUrlKey()
+    {
+        return $this->getData(self::URL_KEY);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setUrlKey($urlKey)
+    {
+        return $this->setData(self::URL_KEY, $urlKey);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getContent()
+    {
+        return $this->getData(self::CONTENT);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setContent($content)
+    {
+        return $this->setData(self::CONTENT, $content);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getIsActive()
+    {
+        return $this->getData(self::IS_ACTIVE) !== null
+            ? (int) $this->getData(self::IS_ACTIVE)
+            : null;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setIsActive($isActive)
+    {
+        return $this->setData(self::IS_ACTIVE, $isActive);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getCreationTime()
+    {
+        return $this->getData(self::CREATION_TIME);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setCreationTime($creationTime)
+    {
+        return $this->setData(self::CREATION_TIME, $creationTime);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getUpdateTime()
+    {
+        return $this->getData(self::UPDATE_TIME);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function setUpdateTime($updateTime)
+    {
+        return $this->setData(self::UPDATE_TIME, $updateTime);
+    }
+}